playground/requirements.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

pytest==6.2.4
googletransit==4.0.0
pandas==1.3.0
streamlit==0.87.0
requests==2.26.0
gtfs-realtime-bindings = 1.0.0
protpbuf
